TOML fot Prompt, Logs of all convos and /ping command to check is it work without waiting
I have changed Json to TOML as this file format is much better for prompting usability. Also added logs history, that you can find in the folder that will be created as soon as you will have the first log. And Added "/ping" command to check fast is it work without a need to go to the Telegram.